requirements.txt
customtkinter
google-genai
python-dotenv
requests
